refactor: extracted dim overlay, panel accent, and share action logic into lib folder
